Update README.md
Experienced Research Assistant with a demonstrated history of working in the higher education industry with a proven record of success in Medical Diagnostic Marketing. Someone who is highly inquisitive and analytical, I am skilled at developing and implementing the methodology, assisting in data collection, and analyzing the results. I am currently looking for opportunities in Healthcare Industries with challenges to develop myself. From that, I shall contribute to the development of your company